+++ src/UtilXml.cs
@@ -24,6 +24,35 @@ namespace Dec
             return root.Elements().Where(child => child.Name.LocalName == name).SingleOrDefaultChecked();
         }
 
+        internal static XElement ElementNamedWithFallback(this XElement root, string name, string inputLine, int lineNumber, string errorPrefix)
+        {
+            var result = ElementNamed(root, name);
+            if (result != null)
+            {
+                // yay
+                return result;
+            }
+
+            // check to see if we have a case-insensitive match
+            result = root.Elements().Where(child => string.Compare(child.Name.LocalName, name, true) == 0).FirstOrDefault();
+            if (result != null)
+            {
+                Dbg.Err($"{inputLine}:{lineNumber}: {errorPrefix}; falling back on `{result.Name.LocalName}`, which is not the right case!");
+                return result;
+            }
+
+            if (root.Elements().Any())
+            {
+                Dbg.Err($"{inputLine}:{lineNumber}: {errorPrefix}; options include [{string.Join(", ", root.Elements().Select(child => child.Name.LocalName))}]");
+                return null;
+            }
+            else
+            {
+                Dbg.Err($"{inputLine}:{lineNumber}: {errorPrefix}; no elements to use");
+                return null;
+            }
+        }
+
         internal static string GetText(this XElement element)
         {
             if (element == null)