+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 * Better text quoting among many error messages.
 * Dec.ToString() now returns type data as well.
 * Better error message if attempting to Record a never-registered Dec.
+* Better error messages and recovery on errors with `key` and `value` elements in <li> Dictionary lists.
 
 ### Fixed
 * Parser.LoadDirectory's prefix filtering breaks if given a directory that includes `.` or `..`.